Patch FindPython3 usage in rocFFT and hipFFT. (#189)
If https://cmake.org/cmake/help/latest/module/FindPython3.html succeeds
(and it must, given the `REQUIRED` argument), the `Python3_EXECUTABLE`
result variable will be set. This should be used instead of a custom
`PYTHON3_EXE` variable. Fixing this allows for these projects to get
further in the configure step on Windows when not using the provided
`toolchain-windows.cmake` files.