Introduce a new configure_ci script. (#241)
Progress on https://github.com/ROCm/TheRock/issues/199. This gives us a
place to configure what workflows should do, based on inputs like the
files changed and labels applied to a pull request. The approach was
inspired by prior work over in
https://github.com/iree-org/iree/blob/main/build_tools/github_actions/configure_ci.py.
Many features there aren't needed here (yet?) though.

## Behavioral changes

For now all this does is skip `build_linux_packages` and
`build_windows_packages` in `ci.yml` if the diff relative to the base
commit only modifies files matching certain patterns (like `*.md`). This
happens for both `push` and `pull_request` triggers but does not happen
for `workflow_dispatch` triggers.

## Testing

I've done some testing on my fork with various pull requests
(https://github.com/ScottTodd/TheRock/actions/workflows/ci.yml) and the
script logs what it does in a few places, but this sort of configuration
and scripting is always a bit tricky. I might try using it with
https://github.com/nektos/act or a test wrapper script that emulates the
environment variables that GitHub sets. We can also add unit tests.

## Future work

The next major feature I'd like to add here is a way to define the
matrix of build/test targets and then a way to opt-in to extra targets
outside of the default set. I think we'll want this sort of script for
that matrix as well as test selection. If this feels over-engineered it
should be easy enough to iterate on a few approaches.