Completes implementation of python rocm-sdk[devel] package. (#253)
* New CLI commands that coordinate with the rocm-sdk[devel] package to
self-install and print the corresponding paths:
  * `rocm-sdk path --cmake`
  * `rocm-sdk path --bin`
  * `rocm-sdk path --root`
* This *should* be sufficient to build frameworks and such, but I have
not actually tested any downstream build on these wheels.
* Frameworks and users will need a new API call added to their
__init__.py which preloads the libraries they depend on (we will add a
`rocm_sdk.loader` for this instead of how CUDA does it where they
scatter random dlopens and paths in the code).
* The result should be that at *build* time, the rocm-sdk[devel] package
will provide a full ROCm installation with all dev tools, etc.
* But at runtime/deployment time, dependent projects need only depend on
the runtime packages they require, which have been managed for size.
* I'm using a discouraged but supported feature of the Python package
ecosystem where the devel package is a self-modifying package. This is
necessary because, unlike the runtime packages, the devel package has
path structures and link farms that cannot be represented in normal
Python packages. Therefore, we treat it as an overlay, distributing an
embedded tarball that is dynamically expanded to augment the development
package when requested. The recommended PyPA accounting of the RECORD is
done, and the result is that `pip uninstall` of the package does in fact
clean up all installation artifacts, whether the embedded development
image has been expanded or not. There are some corner cases that I will
make more robust in the future.
* Now that I have the POC and know exactly what it needs to do, I will
be rewriting linux_python_dist_split.py for better scalability (i.e. so
that we can easily support more libraries and have less fragility). The
symlink handling specifically is getting quite thick and needs a
rewrite. But the tests pass and are fairly comprehensive in terms of
what can go wrong from a packaging perspective. That will aid the
rewrite.