Add THEROCK_SPLIT_DEBUG_INFO and do split .build-id based debug files if enabled. (#305)
* Linux only.
* If enabled, adds staging install time steps to populate the stage
install `.build-id/` directory tree.
* All sub-projects will be augmented to split their debugging files in
this way, and the staging install will be done with `--strip`.
* Verified it works by starting gdb and pointing at the debug directory
(`set debug-file-directory build/dist/rocm`) and debugging clang (`file
build/dist/rocm/lib/llvm/bin/clang`).
* The artifact specific `.build-id` trees can be uploaded to OS specific
repositories, served through a symbol server, or hosted on S3 (and gdb
can be configured to fetch them by URL as needed).
* Before being usable, we need to tweak debug levels project wide to
balance utility with size. And we need to prune the LLVM installation
overall: we shouldn't even be building most of the things there that are
adding a lot of symbol bloat (~16GiB currently, and the rest is a
rounding error compared to that).