Add libdrm patch to inline device ids. (#287)
* Upstream chose to store the device ids at a static location on the
file system. While this perhaps has a use case in a long term support OS
situation, it is fine for us to just embed the id/name associations.
* We need relocatable packages, so eliminating this filesystem lookup is
helpful.
* This just replaces the amdgpu_asic_id.c with an inline lookup. The ids
file is still copied to the output (didn't patch the meson build to
exclude it).
* Verified with `rocm-smi -a` on my GPU and also manually hacked the
device id list to change my device name (and verified that this gets
reported).

Fixes #147

---------

Co-authored-by: Marius Brehler <marius.brehler@amd.com>