Refactor the python packager. (#307)
* Extracts out mechanical code to classes in py_packaging.py. This also
gives us a place to keep some additional accounting which makes some of
the code simpler.
* Makes the packager dynamically query the _dist_info.py file that is
interpreted at build/run-time, keeping everything operating on the same
source of truth of packaging relationships.
* Sets the code up so that we can easily declare dependencies between
the packages and support more fine-grained library packages in the
future.
* When built with current artifacts, this exposes a number of
RPATH/linking issues that are latent:
  * rocprofiler depends on aqlprofile, which is still hard to come by.
* Some RPATHs are messed up in suite-sparse, causing libcholmod and
libamd to not be found in some tests.
* Some libraries just don't appear to be loadable (profiler) and need
more analysis.