Drop rocprofiler-register on Windows. (#325)
Cherry-pick of https://github.com/ROCm/TheRock/pull/316, since it's a
generally useful change and the original branch on
https://github.com/ROCm/TheRock/pull/314 is stalled working through
other issues. I also included an additional documentation change.

Maintaining this patch is getting harder as more and more Linux-specific
code is getting added, so instead drop the subproject entirely on
Windows instead of trying to gracefully make it no-op. This will push
the responsibility for checking for this dep onto other subprojects
throughout the build system.

Closes https://github.com/ROCm/TheRock/issues/301.