Refactor CI/CD workflows to support a matrix of targets (#310)
Still want some questions answered so WIP, but testing to make sure
labels works fully through

Adding CI/CD support for a matrix of targets! Have a few questions, but
I can show this works with ~ labels ~

- For workflow dispatch, users will need to provide a comma-separated
list of targets. Feel free to take a look at the inputs/results of this
[workflow dispatch
test](https://github.com/ROCm/TheRock/actions/runs/14118446514)
- For PRs, we are using labels to specify which OS + machine to run,
running on this
[run](https://github.com/ROCm/TheRock/actions/runs/14118836647)...
checks are being weird and hiding at the bottom even though I see it at
the top of this PR

Progress on #221 

Next todos (by priority):
- Integrating with nightly release and integrating those workflows with
this
- Expanding the matrix
- Solving PR trigger (labels? comments?)