Skip rocprofiler-sdk on unsupported Windows platform. (#383)
Progress on https://github.com/ROCm/TheRock/issues/36.

This enables configuring with `-DTHEROCK_ENABLE_BLAS=ON` on Windows and
building of hipBLAS-common.

Per
https://rocm.docs.amd.com/projects/install-on-windows/en/latest/reference/component-support.html,
the
[ROCProfiler](https://rocm.docs.amd.com/projects/rocprofiler/en/latest/)
program is not supported on Windows today, and users are directed to use
[Radeon GPU Profiler](https://gpuopen.com/rgp/) instead. Additionally,
the roctracer project reports a fatal error during configuration on
Windows:
https://github.com/ROCm/roctracer/blob/a847d331ff2886d23fdb7946695df7fe440c5f57/CMakeLists.txt#L38-L41
```cmake
## Build is not supported on Windows plaform
if(WIN32)
  message(FATAL_ERROR "Windows build is not supported.")
endif()
```

---------

Co-authored-by: Marius Brehler <marius.brehler@amd.com>