Enable windows bundled sysdeps. (#398)
* Enabled the bundled sysdeps mode by default on Windows, adding the
sqlite3 (which was already in common) and zlib libraries.
* I believe this fixes the root cause from #375 without requiring an
LLVM patch:
* LLVM was expecting the CMake ZLIB package to advertise a loose
ZLIB_LIBRARY and ZLIB_INCLUDE_DIRS, but our version was not. This was
causing the check_symbol_exists() to fail, even though the package was
found.
* Includes the zlib FORCE_ON change from #342 and should supercede the
need for that patch.
* Also ports the zlib patch script from bash->python and uses it from
both the linux and windows build.
* May need another tweak or two on the Linux side: since I developed
this on Windows, I was just eyeballing Linux.

Fixes #375
Supersedes #342