Fix the bundled boost build to configure properly on Windows. (#440)
* Uses a hack to configure the desired libraries (since the
bootstrap.bat is broken and does not parse this).
* Disables x32 builds on Windows.
* Only builds release.
* Uses the b2 install command. This is *really* unforunate but lets us
avoid a fragile symlink situation. It has to materialize ~17k headers
and will do so regardless of what was asked for. This extends the build
time on Windows by dozens of seconds because serial filesystem access at
that scale is slow. However, hopefully this overlaps with some other
build steps or can be cached.