Adds a magic env var TEATIME_FORCE_INTERACTIVE and sets it to 0 in CI jobs. (#494)
This causes subproject logs to only be output to the console on failure
vs success. Given the state of warning spew in the projects and how hard
it is to track through running console logs for this stuff, this is the
better setting. All detailed triage needs to happen on log files, which
we persist in S3.

Some third-party subprojects which we don't control were already setting
this at the project level. Setting at the CI level just lets us exert
special control project wise.

Also switches the Linux output to be prefixed by project name instead of
fancy expando handles. I've found this to be more ergonomic, especially
in non-interactive mode.

Also removes the watch_top_processes thing on Linux. We have better ways
to do this now.

Normal users remain subject to the full fury of the warnings.