Adding rsync to provision script (#465)
Adding `rsync` to provide the ability for local code to be moved to test
machines. This is just a working version, but it will be paired with
TheRock testing suite and will be tweaked during TheRock testing suite
implementation.

This script can be used by local developers wanting to get a specific
release version/CI run, CI runners and eventually test machines.
Basically, test suite will use this script to install rocm to test
machines

Progress on #459