[Windows] Install more deps on CI runners. (#584)
Progress on https://github.com/ROCm/TheRock/issues/543. These should be
all the deps needed to build the rest of the subprojects, but builds are
stalling at some point. I'd like to checkpoint while we continue
debugging.

Installing strawberry perl (including gfortran) also puts gcc on the
PATH, and that sometimes gets picked up as the compiler for CMake/Ninja
to use at both the top level and in subprojects that aren't explicit
about inheriting the compiler options from the superproject.