[windows] Add BLAS test/bench binaries to dist. (#591)
Fixes https://github.com/ROCm/TheRock/issues/522.

The executable files have the `.exe` suffix on Windows, so add a
trailing wildcard in the glob patterns. We could also teach
[`fileset_tool.py`](https://github.com/ROCm/TheRock/blob/main/build_tools/fileset_tool.py)
to try the platform-specific suffix if the file is missing a suffix.

Maybe we could also handle `lib[name].so` -> `[name].dll` and other
similar transforms, but I _do not_ see Windows equivalents for
`lib/libhipsolver_fortran.so` or `lib/libhipblas_fortran.so`.