[torch][windows] Apply fixes for PyTorch+ROCm on Windows. (#647)
Progress on https://github.com/ROCm/TheRock/issues/589. Fixes
https://github.com/ROCm/TheRock/issues/590. Fixes
https://github.com/ROCm/TheRock/issues/627.

This is a revised version of a patch from @jammm, made in collaboration
with @scottt as well. The diff between the original patch and what I am
sending here can be viewed at
https://github.com/ScottTodd/TheRock/blob/pytorch-windows-cleanup-wip/external-builds/pytorch/patches/pytorch/v2.7.0/pytorch/hipified/0014-Cleanup-after-testing-with-clang-cl.patch.
Summarizing the delta:

* Revert some changes that were not required in my testing
* Trim some MSVC handling, in favor of focusing on just clang-cl first
* Address hipcc warnings

With this patch, I've been able to build PyTorch wheels with ROCm
support on Windows using
[`build_pytorch_windows.sh`](https://github.com/ROCm/TheRock/blob/main/external-builds/pytorch/build_pytorch_windows.sh),
though there are still some remaining changes needed for a good out of
the box experience (see
https://gist.github.com/ScottTodd/cb4f9f5a697cce91084977ed3674a4bd and
https://github.com/ROCm/TheRock/discussions/409). The script at
https://github.com/ROCm/TheRock/blob/main/external-builds/pytorch/windows_patch_fat_wheel.py
can also be used to produce a "fat wheel" that bundles ROCm instead of
relying on a separate install.

I intend to send the patch here to upstream
https://github.com/pytorch/pytorch as well, perhaps split into a few
smaller PRs.

Co-authored-by: Aaryaman Vasishta <jem456.vasishta@gmail.com>
Co-authored-by: Scott Todd <scott.todd0@gmail.com>