Allow to build components from `rocm-libraries` (#682)
This adds the `THEROCK_USE_EXTERNAL_ROCM_LIBRARIES` option to allow
consuming source from the `rocm-libraries` monorepo instead of a
submodule. The option must be used together with
`THEROCK_ROCM_LIBRARIES_SOURCE_DIR` to point to the monorepo (relative
and absolute paths are supported). With this, the monorepo can either be
a submodule in TheRock itself or can be an checkout outside of TheRock.

For now, this is enabled for rocPRIM.