Bump rocprofiler-sdk to HEAD, rebase patches, and add aqlprofile. (#747)
* rocprofiler-register and rocprofiler-sdk switched to `amd-staging`
(from mainline) and updated for the first time in a few months, now that
we are looking at profile support again.
* Adds the newly open sourced `aqlprofile` and wires it to
rocprofiler-sdk, which allows for an open source build of the stack.
* Switches some base jobs to BACKGROUND_BUILD.
* Extends subproject support so that INTERFACE_LIBRARY_DIRS propagates
to CMAKE_LIBRARY_PATH, allowing `find_library()` to prefer super-project
dependent projects.
* Rebases and combines some small rocprofiler-sdk build fixes, dropping
those that are no longer needed.
* Temporarily reverts
https://github.com/ROCm/rocprofiler-sdk/commit/fec74141b060b75afbcf41077796c2380dfc7198,
disabling ABI compatibility with ROCm 6.4 until underlying HIP build
breaks are resolved.
* Removes a couple of dead links.

This is sufficient to make rocprofv3 build in a purely OSS setting.
There is still a small binary-only decoder library that we need to
download and include, and I need to add roctracer and do some fixups to
get the whole stack functioning. Will do that next.