[python] Plumb through distribution info from the build. (#769)
* First steps towards building split-gfx-arch packages by supporting
"dist" targets (the superset of targets that the overall distribution is
built for) and build targets (the subset of dist targets that we are
building as part of this shard).
* Adds new `USE_DIST_AMDGPU_TARGETS` project tag which has it populate
the GPU_TARGETS from the dist targets and applies this to all target
neutral projects.
* Populates `GPU_TARGETS` and `AMDGPU_TARGETS` regardless of the
selected toolchain. Prior, these were only being populated for the hip
toolchain, which was causing some inconsistencies when building runtime
targets.
* Also adds `DISABLE_AMDGPU_TARGETS` and applies it to aqlprofile
because of a bug (see https://github.com/ROCm/aqlprofile/issues/2).
* Works around a nasty bug that was encountered while testing and has
probably been ghosting users for a while:
https://github.com/ROCm/rocRAND/issues/653
* Adds `rocm-sdk version` (and test) which prints the version.
* Adds `rocm-sdk targets` (and test) which prints the dist targets that
it was built with.
* Bumps the major version of rocprofiler-sdk-roctx to 1 (TODO: need to
probe all of these versus hard-coding).
* Enables `rocm-sdk test` coverage for the profiler libraries now that
aqlprofile is available/works.
* Installs `share/therock/dist_info.json` as part of every build, giving
us a place to emit various distribution metadata.

Tested: Local build of ROCm plus python packages, install and `rocm-sdk
test`.

Progress on #703