Add index generation for Pytorch wheels (#777)
This PR adds:

1. The index generation for Pytorch wheels
2. Fix "+" problem in the wheel key link that is omitted in the indexer
3. Improved the modularity of the design with the input parameters

For instance:
"+" sign in the key replaced with the space which causes the error
below:
<Key>gfx94X-dcgpu/torch-2.7.0a0
rocmsdk20250530-cp312-cp312-linux_x86_64.whl</Key>

Correct one: 

<Key>gfx94X-dcgpu/torch-2.7.0a0+rocmsdk20250530-cp312-cp312-linux_x86_64.whl</Key>

The error seen in the browser:
```
This XML file does not appear to have any style information associated with it. The document tree is shown below.
<Error>
<Code>NoSuchKey</Code>
<Message>The specified key does not exist.</Message>
<Key>gfx94X-dcgpu/torch-2.7.0a0 rocmsdk20250530-cp312-cp312-linux_x86_64.whl</Key>
<RequestId>18YCBF52GCKJ14RG</RequestId>
<HostId>45PdRkbM0BsN8/xUq5wj+D9kBFTmn6fu2/F2jbXG8j5BSLU4RPmSAdVCiRFz3EXswfWA/VQwces=</HostId>
</Error>
```

The wheels can we downloaded now from index file:
The index file needs to be refreshed in browser.

https://therock-nightly-python.s3.us-east-2.amazonaws.com/gfx94X-dcgpu/index.html

---------

Co-authored-by: Marius Brehler <marius.brehler@amd.com>