Add ArtifactName.from_filename util function. (#838)
Splitting this change off from https://github.com/ROCm/TheRock/pull/772.

This will make it easier to work with artifacts that are on S3 and other
non-local locations, where we'll have a URL and a filename, not a Path.
We can also see about using
[`cloudpathlib`](https://cloudpathlib.drivendata.org/stable/), "A Python
library with classes that mimic `pathlib.Path`'s interface for URIs from
different cloud storage services".