Adds an `export_source_archive.py` script. (#889)
This script is intended to produce canonical source and prebuilt
archives for the project, which are inputs to downstream native
packaging flows (like DEB, RPM, etc).

As an example, this script can produce archives that serve the following
roles:

* Used as a Debian `.orig.tar.gz` upstream source archive.
* Included in a RedHat SRPM.
* Archived on release pages to preserve exact released sources with
everything included to build the project at a point in time.

While most vendors will choose to perform their own native build against
pure source archives, we also support a mode where we generate "prebuilt
source archives". These archives are structurally similar but include
pre-built artifacts from a prior build system along with a skeleton of
enough sources to re-assemble/package the project. The intent is that
these can feed into packaging flows, causing them to skip their local
build in favor of simply using pre-built artifacts. We intend to use
this for cheaply generating our own vendor packages for a variety of
operating systems by building on our common feeder distributions that
are compatible with a wide fanout of systems. The same packaging spec
files should be able to work with minimal modifications from either
(since the PREBUILT variety simply causes the packaging rules to *not*
build anything as it usually would).

Currently a pure source archive is about ~830MiB, largely driven by the
size of LLVM and some large artifacts in the kernel library projects
(this can be reduced to ~700MiB with higher compression, at the expense
of much longer export time). This will vary a bit over time as we have
plans to eliminate some such bloat, but surely other things will take
their place.