[ccache] Add setup_ccache.py helper script and tune the build. (#893)
Because TheRock bootstraps a compiler as part of its build, we have to
take special care to teach ccache how to determine that the compiler in
each build is the same and should not invalidate the caches of what is
built with it. ccache has built in options that don't quite suite, so we
use a custom compiler check which implements a two stage full-content
based check of the compiler binary and dependent libraries. Stock CMake
can verify the content but it does not check all dependent libraries,
which doesn't work for us, since most of the compiler is in libLLVM.so
and libClang.so. Since the options get very fiddly, I scripted a one
stop ccache configuration script that is suitable for use in CI and by
users. This will likely need some more tuning in the future so it is
better to have it scripted vs as loose instructions.

After a lot of pouring through logs, I did conclude that I have to set
`export CCACHE_SLOPPINESS=include_file_ctime` to achieve full cache
hits. I left a note in the code as to why this is. Raised upstream
discussion: https://github.com/ccache/ccache/discussions/1603

With this, we should be able to set up a ccache backend in k8s and use
that across the board.

These tests were done on ccache 4.11.3, but it should be valid for any
version >= 4.11 (specifically,
https://github.com/ccache/ccache/pull/1571 is needed to properly handle
device offload compression). Note that this version of ccache has not
made it into any distros that I am aware of and needs to be installed
manually.

Testing on gfx1100:
* Before: ~45m regardless of cache state (because ccache was failing so
many things).
* After: 11m (rccl becomes the long tail because it does non-cacheable
LTO)
* Without rccl: 9m (more than half of this time is just due to needless
serialization and poor organization of projects in the BLAS stack)
* Note that rocfft does direct RTC compilation, which is non cacheable,
so it takes a constant ~2m of high parallelism per arch no matter what.

There are still a number of nit picky things failing to cache, but they
do not seem to implicate the critical path much. The biggest outlier is
that something that rocprofiler-sdk is doing is making it almost
entirely uncacheable. I collected logs but didn't get to the long tail
yet.

---------

Co-authored-by: Scott Todd <scott.todd0@gmail.com>