[pytorch] Add build support for the vendored pytorch-triton-rocm wheel. (#948)
* Merely adds scripting support for checking out the proper pinned
repository and building if done.
* Should be NFC on any existing build pipelines until enabled.
* The build procedure is a bit annoying because it relies on a couple of
action at a distance build flags and pins that need to line up exactly.

As part of this, I did a code review of what the Triton build is doing
behind the scenes, some of which is problematic. I have looped the
Triton maintainer in and we will fix in followups:

* The AMD backend has bugs in terms of how it is finding the
libamdhip64.so having to do with it using dlopen to probe existing
loaded libraries by barename (i.e. `libamdhip64.so`) vs soname (i.e.
`libamdhip64.so.6`), resulting in some exotic fallbacks to scanning
process tables that should not be necessary. The entire procedure is non
deterministic and I can imagine ways it can fail in the field.
* We need to add a probe for installation via the ROCM wheels and use
the API it provides to find the libraries vs relying on heuristics.
* By default the Triton build fetches pre-built binaries for LLVM static
libraries it is pinned to. These appear to be built on manylinux2014 vs
newer glibc. Static libraries can have subtle incompatibilities across
gcc-toolset versions. Appears to not be a problem now but worth keeping
an eye on.
* There is a lot of stuff in the pytorch/AMD scripts for bundling shared
libraries and rewriting rpaths. None of this appears to be currently
load bearing in present versions of Triton and was left out. In
addition, even if it were, the rocm wheels would take care of it vs
requiring exotic bundling.

We'll fix all of this going forward once the basic build machinery is in
place and functioning.

---------

Co-authored-by: Scott Todd <scott.todd0@gmail.com>