Add Pytorch tests into `test_pytorch_wheels.yml` workflow. (#903)
Added a workflow called` test_pytorch_wheels.yml` that runs the Pytorch
tests below.
The tests are not fully passed because of multiple fails and errors on
test cases. Eliminating those errors is in a progress currently and the
goal is pass all of the tests to make workflow finish with the passing
statistics.

The type of the errors we have:
1) TypeError
2) RuntimeError
3) AssertionError
4) AttributeError

The Pytorch tests are added:
```
 test_nn.py
 test_torch.py
 test_cuda.py
 test_unary_ufuncs.py
 test_binary_ufuncs.py
 test_autograd.py
```
TODO: Add back `test_torchinductor.py` and `test_ops.py` when the
AttributeError solved.

Temporary used `--no-hipify --no-patch` for python
`external-builds/pytorch/pytorch_torch_repo.py checkout`. This will be
removed after the PR is landed: https://github.com/ROCm/TheRock/pull/928

Temporary used `-k` for
https://github.com/ROCm/TheRock/pull/903/files#diff-a707520c93985455d9e178b2f548a8f691704d58bcf5e314b469ff6553847b58
After related errors are solved, this flag will be removed. 

The current test stats:
https://github.com/ROCm/TheRock/actions/runs/15990272172/job/45102132701
70 failed, 37802 passed, 2365 skipped, 2 deselected, 53 xfailed, 2
errors
FYI, @stellaraccident @amd-chrissosa