Explicitly set the ROCm version to use for building torch (#965)
Instead of pulling the latest ROCm packages, this adds the option to
pass the ROCm version that should be installed. This is also used to
determine the local version identifier (which defaults to `rocmsdk` and
the local date). Packages created with the modified workflow have the
following version identier:

* Nightly builds: torch-{torch base version}+rocm{base rocm
version}rc{release date}
* Dev builds: torch-{torch base version}+rocm{base rocm version}.dev{dev
version}.{rocm commit sha}

Examples for packages are:
*
torch-2.7.0a0+rocm7.0.0.dev0.515115ea2cb85a0b71b5507ce56a627d14c7ae73-cp312-cp312-linux_x86_64.whl
* torch-2.7.0a0+rocm7.0.0rc20250702-cp312-cp312-linux_x86_64.whl

Noticeably, the local version identifier starts with `rocm` and not with
`rocmsdk` as in the default. ROCm dev builds themselves have a local
version identifier making it necessary to replace the `+` as one would
otherwise have two local version identifiers, resulting in an invalid
version.

The only downside of the proposed local version identifier is that the
ones used for the dev packages result in a version that is evaluated
larger as the ones of the nightly packages. With this, dev packages and
nightly releases cannot be easily mixed and we might want to revise this
later.