Trigger building PyTorch wheels via portable linux packages workflow (#983)
Till now, the `release_linux_pytorch_wheels.yml` workflow was run on
schedule. As a consequence `build_linux_pytorch_wheels.yml` picked up
the latest available ROCm Python packages, not necessarily matching the
version number later encoded in the PyTorch wheel.

This is revised and instead of building PyTorch wheels on schedule, the
build is triggered from the workflow used to build the ROCm Python
packages. To make sure the correct nightly or dev version is used, or
rather pip installed, the explicit version is passed by the triggering
workflow.
As a side effect this also enables building PyTorch wheels for gfx950.

We may want to further revise this later, as there is no clear and easy
to follow link between the triggering and the triggered workflow.

Closes #923