Wire up testing PyTorch wheels (#998)
This wires up testing the PyTorch wheels build in the
`build_linux_pytorch_wheels` workflow. The package version to test is
grabbed from the package name after building it and passed to the next
job. This could probably made more robust by moving it to a script but
should be acceptable for now. At the moment tests get only triggered for
the platforms that are also used to test TheRock nightlies, which for
now only is gfx942.

Furthermore, this adds another PyTorch test to the skip tests (here
`test_index_add_correctness`), which fails with `AssertionError:
Tensor-likes are not close!` on gfx942 using Python 3.11 but passes for
3.12 and 3.13.