Extend pytorch build to work with our official 2.7 backport. (#1027)
This enables us to build off of the ROCm/pytorch release/2.7 branch,
enabling 2.7.1 on ROCm 7.

jeffdaily and I cherry-picked the couple of missing patches from HEAD to
make this configuration work with TheRock.

When building with the pre-built aotriton, it hits against #1025.

Should be NFC for all other builds.

Progress on #1026.

---------

Co-authored-by: Marius Brehler <marius.brehler@amd.com>