Sets up CMAKE_PREFIX_PATH so that it can find all super-project packages. (#1046)
* Adds a debug flag `THEROCK_USE_SAFE_DEPENDENCY_PROVIDER` which allows
the safe dependency provider to be disabled. This allows debugging that
the project is clean with pure find_package semantics.
* For each sub-project, creates a `prefix` directory and adds it to
CMAKE_PREFIX_PATH. This prefix directory contains trampoline package
config files that are generated to do hard-coded redirects to those at
known locations in the sub-projects.
* Uses the new debug flag to sanitize the project for these cases:
* ROCR-Runtime has some shady finders for LibElf that the dependency
provider was handling. Adds a pre-hook to prime the LibElf package
explicitly, which is safe/consistent in both modes.
* There are still some issues in this mode further down project, but
these can be fixed later (i.e. it catches a comgr version mismatch that
is legitimate).
* This should make sub-cmake invocations work properly from a
find_package perspective so long as CMAKE_PREFIX_PATH is propagated
(which is pretty standard practice).