Stabilize compiler version fingerprint. (#1095)
* On checkout, if submodules are patched, a sibling file `.{name}.smrev`
is written containing a line for the origin repository and a mangled
revision (indicating the base revision of the repo and a digest of all
raw patches applied).
* During build of LLVM, if the .smrev file is present, it is read and
LLVM_FORCE_VC_REPOSITORY and LLVM_FORCE_VC_REVISION are set with the
hard-coded contents.
* This overrides the auto fingerprinting logic in the LLVM build,
resulting in a stable origin stamp regardless of whether patches were
applied or not.
* This appears to be enough to produce a bit for bit identical compiler,
allowing existing ccache logic to do the right thing. In the future,
however, for CI, we could change the cache logic to look at the first
line of `$COMPILER --version`. This however does not capture all things
that can indicate a different compiler, so we should try the
reproducible build as is first.