Enable pytorch nightly builds and fix the triton versioning problem (#1208)
Two changes were needed in this commit to be able to do both the nightly and release builds.

1) Enable nightly builds for pytorch, pytorch vision, pytorch audio and triton

2) Fix the versioning problem on Triton nightly builds which showed up due to the different
    ways how the version number was constructed in the setup.py on triton/rocm/release/internal/3.3.x builds compared to upstream Triton

Full Triton version is constructed at the moment in upstream Triton's setup.py in a following way:
    
    TRITON_VERSION = "3.4.0" + get_git_version_suffix() + os.environ.get("TRITON_WHEEL_VERSION_SUFFIX", "")

While the Triton release builds construct the version in this way:

    TRITON_VERSION = "3.4.0" + os.environ.get("TRITON_WHEEL_VERSION_SUFFIX", "")
    
Problem here is that Triton version can only contain one plus-character. The plus-character
is used for distinguishing additional version information from the base major.minor.patch version scheme.
    
If the Triton build is made from the release branch, then get_git_version_suffix() will return "". In every other case it will return "+githash".
    
This means that if the build is done from the release-branch and TRITON_WHEEL_VERSION_SUFFIX environment variable is used, then the env[TRITON_WHEEL_VERSION_SUFFIX] should start with the
plus-character.

In other cases the env[TRITON_WHEEL_VERSION_SUFFIX] should start with the dash-character.

If implemented in the CI script side, it adds easily an extra complexity to the build scripts that can be handled more easily by implementing a fix to Triton's setup.py to ensure it will work in both use cases.

After implementing a various different ways to handle the problem, we came to conclusion that best way is to manage this as an patch that will send later to upstream Triton for review in hope to get it merged there. Suggested patch will only allow 1 plus character in the version suffix part of the version string and if needed convert other "+"characters to "-"-character.

---------

Signed-off-by: Mika Laitio <mika.laitio@amd.com>
Co-authored-by: Marius Brehler <marius.brehler@amd.com>