Add a prebuilts.py CLI script for managing project enable/disable.
This provides a user friendly way to enable which subset of the projects are actively enabled for build, allowing the user to take fine grained control of which parts of the build graph get actively rebuilt and which parts are effectively frozen.

Future enhancements will support import/export of build tree state (ie. for boostrapping without building all components) and artifact download from CI.

See file comments for usage and description.